requirements.txt
psutil~=5.9.8
py-cpuinfo~=9.0.0